Start implementation of reporting feature where latest value gets stored and in addition to that, the measured values are persisted per day.

+++ source/RaspiAir.Measurement.Services/MeasurementService.cs
@@ -52,7 +52,7 @@ internal class MeasurementService : IBackgroundService
             new HumidityChangedEvent(value, DateTimeOffset.UtcNow));
     }
 
-    private void HandleCo2ConcentrationChanged(double value)
+    private void HandleCo2ConcentrationChanged(int value)
     {
         this.eventBroker.Publish(
             new Co2ConcentrationChangedEvent(value, DateTimeOffset.UtcNow));
